==> 06_TBPCS_output_v1.2.Rmd <==
---
title: "TB Patient Cost Survey with dummy data"
author: ##### "Andrea Pantoja / Nobu Nishikiori"
date: ##### "`r format(Sys.time(), '%d %B %Y')`"
output:
  html_document:
    fig_caption: yes
    fig_width: 9
    toc: yes
    toc_depth: 4
    toc_float: yes
  word_document:
    toc: yes
    toc_depth: '4'
  pdf_document:
    toc: yes
    toc_depth: '4'
  powerpoint_presentation:
    toc: no
editor_options: 
  chunk_output_type: console
---
\      

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(magrittr)
library(haven) # for read_dta() 
library(labelled) # for labelling
library(survey)
library(binom)
library(finalfit)
library(knitr)
library(kableExtra)

#  
setwd <- function(dir) {
  if (missing(dir) || is.null(dir) || dir == "") {
    dir <- "C:/Users/Takuya"
  }
  base::setwd(dir)
}

setwd()

setwd("./Dropbox/10.GTB/Working folder Taku-Nobu/R_generic_tzn_based/")
output_loc <- "out_iso3/"
load("C:/Users/Takuya/Dropbox/10.GTB/Working folder Taku-Nobu/R_generic_tzn_based/R_data/TBPCS_iso3_complete.RData")
## load functions ####
source("funs.R")

options(scipen=10)

```


***
\pagebreak
    "tb_type","treat_group","phase","hiv","self_admin"
   
#### Table 1. Descriptive statistics of survey sample by drug-resistance status   
```{r Table 2, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 2 ####
## Tabulation by using 
## finalfit::summary_factorlist 
##
dependent = "mdr"
expl_1 = c("sex","age","age_c", "region", "insurance", "educ_level", "employ_before", "emp_main", "below_poverty", "hhsize" )#,"treat_group_nmdr","treat_group_mdr")
expl_2 = c("income_hh_pre_re","expend_hh")
expl_3 = c("treat_group","tb_type","hiv","place_diag","facility_type","phase","duration_int","duration_cont","self_admin_current_int","self_admin_current_con","current_hosp","prev_hosp","t_stay_current")

Table2_1 <- 
  all %>% 
  mutate(
    age = ff_label(age, "Age (years)"),
    age_c = ff_label(age_c, "Age group"),
    sex = ff_label(sex, "Sex"),
    region = ff_label(region, "Region"),
    insurance = ff_label(insurance, "Insurance status"),
    educ_level = ff_label(educ_level, "Education level"),
    employ_before = ff_label(employ_before, "Employment status (before having TB)"),
    emp_main = ff_label(emp_main, "Patient is breadwinner in household"),
    below_poverty = ff_label(below_poverty, "Below poverty line (global definition)"),
    hhsize = ff_label(hhsize, "Household size")) %>% 
  summary_factorlist(dependent,expl_1,cont = "mean",cont_range=TRUE,p=T,total_col = TRUE,
                     na_include=TRUE,na_to_missing=TRUE,digits = c(1, 1, 3, 1))
Table2_2_mean <- 
  all %>% 
  summary_factorlist(dependent,expl_2,cont = "mean",p=T,total_col = TRUE,digits = c(0, 0, 3, 0))

Table2_2_median <- 
  all %>% 
  summary_factorlist(dependent,expl_2,cont = "median",cont_range=TRUE,p=T,total_col = TRUE,
                     digits = c(0, 0, 3, 0)) %>% 
  mutate(label="")

table_2_2 <- rbind(Table2_2_mean[1,],Table2_2_median[1,],
                 Table2_2_mean[2,],Table2_2_median[2,]#,
                 #Table2_2_mean[3,],Table2_2_median[3,]
                 )
# Table2_3 <- 
#   d %>% 
# #  mutate(below_poverty_tshs = ff_label(below_poverty_tshs, "Below poverty line")) %>% 
#   summary_factorlist(dependent,expl_3,p=FALSE,total_col = TRUE,na_include=TRUE,na_to_missing=TRUE,
#                      digits = c(1, 1, 3, 1))

Table2 <- rbind(Table2_1,table_2_2#,Table2_3
                ) %>% remove_rownames() %>% dplyr::rename("Variable"=1, " "=2)

kable(Table2, "html", #caption = "Table 2. Descriptive statistics of survey sample by drug-resistance status",
      align = 'llccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %T>%
  cat(., file = paste0(output_loc,"Table 1. Descriptive statistics.html"))

##### clinical chars #####
all$duration_int <- as.numeric(all$duration_int)


Table2_3 <- 
  all %>% 
  mutate(
    treat_group = ff_label(treat_group, "Treatment group"),
    hiv = ff_label(hiv, "HIV status"),
    tb_type = ff_label(tb_type, "Type of TB"),
    place_diag = ff_label(place_diag, "Place of TB diagnosis"),
    facility_diag = ff_label(place_diag, "Place of TB treatment"),
    phase = ff_label(phase, "TB treatment phase"),
    duration_int = ff_label(duration_int, "Duration of intensive phase (month)"),
    duration_cont = ff_label(duration_cont, "Duration of continuation phase (month)"),
    self_admin_current_int = ff_label(self_admin_current_int, "Mode of supervision for TB treatment: intensive phase"),
    self_admin_current_con = ff_label(self_admin_current_con, "Mode of supervision for TB treatment: continuation phase"),
    current_hosp = ff_label(current_hosp, "Currently hospitalized"),
    prev_hosp = ff_label(prev_hosp, "Previously hospitalized in the current phase"),
    emp_main = ff_label(emp_main, "# of days hospitalized in the current phase"),
    t_stay_current = ff_label(t_stay_current, "Below poverty line (global definition)")
    ) %>% 
  summary_factorlist(dependent,expl_3,cont = "mean",cont_range=TRUE,p=FALSE,total_col = TRUE,
                     na_include=TRUE,na_to_missing=TRUE,digits = c(1, 1, 3, 1))

Table3 <- Table2_3 %>% remove_rownames() %>% dplyr::rename("Variable"=1, " "=2)

kable(Table3, "html", #caption = "Table 2. Descriptive statistics of survey sample by drug-resistance status",
      align = 'llccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %T>%
  cat(., file = paste0(output_loc,"Table 2. Clinical TB statistics.html"))

```
\     
  
***
\pagebreak
\    
  
#### Table 3. Model of care for survey sample  
```{r Table 3, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 3 ####
## Tabulation by using 
## finafit::summary_factorlist 
##

dependent = "mdr"
expl_1 = c("current_hosp","prev_hosp","t_stay_current")
expl_2 = c("visit_tot","n_dot_visits","s_fu","n_pickup_total","visit_before")
expl_3 = c("duration_int","duration_cont")
expl_4 = c("weeks_before_tx","delay")

Table3_1 <-   
  all %>% 
  summary_factorlist(dependent,expl_1,cont = "mean",cont_range=TRUE,p=FALSE, total_col = TRUE,
                     na_include=TRUE,na_to_missing=TRUE,digits = c(1, 1, 3, 1)) %>% 
  ff_remove_ref() %>% mutate(levels=ifelse(levels!="Mean (SD)","n (%)",levels))

Table3_2 <- all %>% 
  summary_cont(dependent,expl_2,p=FALSE,total_col = TRUE,digits = c(1, 1, 3, 0))

Table3_3 <- all %>%   
  mutate(
    duration_int = ff_label(duration_int,"Intensive phase (month)"),          # label change needed only for this table
    duration_cont = ff_label(duration_cont,"Continuation phase (month)")) %>% # label change needed only for this table 
  summary_factorlist(dependent,expl_3,cont_cut=3,cont="mean",p=FALSE,total_col = TRUE,digits = c(1, 1, 3, 0))

Table3_4 <- 
  summary_factorlist(all, dependent,expl_4,p=FALSE,total_col = TRUE,digits = c(1, 1, 3, 0))

Table3 <- rbind(Table3_1,Table3_2,Table3_3,Table3_4) %>% remove_rownames() %>% 
  dplyr::rename(" "=1," "=2)
kable(Table3, "html", #caption = "Table 3. Model of care for survey sample",
      align = 'llccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  pack_rows("Hospitalisation",1,3) %>% 
  pack_rows("Ambulatory care",4,13) %>% 
  pack_rows("Treatment duration",14,15) %>% 
  pack_rows("Treatment delay",16,18) %>%
  cat(., file = paste0(output_loc,"Table 3. Model of care for survey sample.html"))




```
\     
  
***
\pagebreak
\    
  
#### Table 4. Hours lost by patient while seeking and accessing care  
```{r Table 4, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 4. Hours lost by seeking and accessing care ####
## Tabulation by using 
##  twoway_stat
##
# Table4_1 <- 
#   rbind(
#     d %>% twoway_stat(phase,mdr,t_before,fun=n_mean_sd_med_iqr),
#     d %>% twoway_stat(phase,mdr,t_current,fun=n_mean_sd_med_iqr)
#   ) %>% ungroup() %>% 
#   separate(5,into=paste0(c("n","mean","median"),"3"), sep='__') %>% 
#   separate(4,into=paste0(c("n","mean","median"),"2"), sep='__') %>% 
#   separate(3,into=paste0(c("n","mean","median"),"1"), sep='__') %>% 
#   slice(c(1,4,5)) %>% select(-level) %>% 
#   mutate(var=c("Prediagnosis","Intensive phase","Continuation phase"))
# names(Table4_1) <- c(" ",rep(c("n","mean (SD)","median (IQR)"),3))

Table4_1 <- 
  rbind(
    all %>% group_by(phase,mdr) %>% summarize_at(vars(t_before,t_current),n_mean_sd_med_iqr), 
    all %>% group_by(phase,mdr="Total") %>% summarize_at(vars(t_before,t_current),n_mean_sd_med_iqr) 
  ) %>% pivot_longer(3:4) %>% pivot_wider(names_from = mdr, values_from = value) %>% 
  separate(5,into=paste0(c("n","mean","median"),"3"), sep='__') %>% 
  separate(4,into=paste0(c("n","mean","median"),"2"), sep='__') %>% 
  separate(3,into=paste0(c("n","mean","median"),"1"), sep='__') %>% 
  ungroup() %>% slice(c(1,2,4)) %>% select(-1) %>% 
  mutate(name=c("Prediagnosis","Intensive phase","Continuation phase"))
names(Table4_1) <- c(" ",rep(c("n","mean (SD)","median (IQR)"),3))

Table4_2 <- 
  rbind(
    d %>% 
      mutate(income_pre_annual=replace_na(income_pre_annual,replace=0)) %>% 
      mutate(income_now_annual=replace_na(income_now_annual,replace=0)) %>% 
      group_by(mdr) %>% 
      summarize_at(vars(income_pre_annual,income_now_annual),n_mean_sd_med_iqr),
    d %>% 
      mutate(income_pre_annual=replace_na(income_pre_annual,replace=0)) %>% 
      mutate(income_now_annual=replace_na(income_now_annual,replace=0)) %>% 
      group_by(mdr=factor("Total")) %>% 
      summarize_at(vars(income_pre_annual,income_now_annual),n_mean_sd_med_iqr)
  ) %>% ungroup() %>% pivot_longer(2:3) %>% pivot_wider(names_from = mdr, values_from = value) %>% 
  separate(4,into=paste0(c("n","mean","median"),"3"), sep='__') %>% 
  separate(3,into=paste0(c("n","mean","median"),"2"), sep='__') %>% 
  separate(2,into=paste0(c("n","mean","median"),"1"), sep='__')
names(Table4_2) <- c(" ",rep(c("n","mean (SD)","median (IQR)"),3))

Table4 <- rbind(Table4_1,Table4_2)

kable(Table4, "html", #caption = "Table 4. Hours lost by patient while seeking and accessing care",
      align = 'rccccccccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  pack_rows("Hours lost by patients",1,3) %>% 
  pack_rows("Individual income reported",4,5) %>% 
  add_header_above(c(" " = 1, "DS-TB" = 3, "DR-TB" = 3, "Total" = 3)) %T>%
  cat(., file = paste0(output_loc,"Table 4. Hours lost by patient while seeking and accessing care.html"))

```
\     
  
***
\pagebreak
\    
  
#### Table 5. Estimated costs borne by households affected by TB  
```{r Table 5, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 5. Estimated costs borne by households affected by TB ####
## Tabulation by using 
##   dplyr
##
cost_cat <- d %>% 
  mutate(cat_medical=cat_before_med+cat_current_med,
         cat_nmedical=cat_before_nmed+cat_current_accomodation+cat_current_travel+
           cat_current_food+cat_current_nutri,
         total_cost_output=cat_medical+cat_nmedical+income_diff) %>% 
  select(mdr,cat_before_med,cat_before_nmed,cat_before_indirect,
         cat_current_med,cat_current_travel,cat_current_accomodation,
         cat_current_food,cat_current_nutri,cat_current_indirect,
         cat_medical,cat_nmedical,cat_indirect,income_diff,
         total_cost_hc,total_cost_output)

var_label(cost_cat) <- list(
  cat_before_med = "Medical",
  cat_before_nmed = "Non-medical",
  cat_before_indirect = "Indirect costs",
  cat_current_med = "Medical",
  cat_current_travel = "Travel",
  cat_current_accomodation = "Accomodation",
  cat_current_food = "Food",
  cat_current_nutri = "Nutrition supplement",
  cat_current_indirect = "Indirect costs",
  cat_medical = "Medical",
  cat_nmedical = "Non-medical",
  cat_indirect = "Indirect costs",
  income_diff = "Indirect costs (income loss)",
  total_cost_hc = "Total cost (human capital apporoach)",
  total_cost_output = "Total cost (output approach)")

cost_cat_hc <- cost_cat %>% select(-income_diff,-total_cost_output)
cost_cat_op <- cost_cat %>% select(-cat_before_indirect,-cat_current_indirect,-cat_indirect,-total_cost_hc)

Table5a_hc <- 
  rbind(
    cost_cat_hc %>% group_by(mdr) %>% summarize_all(mean_sd_med_iqr),
    cost_cat_hc %>% group_by(mdr="Total") %>% summarize_all(mean_sd_med_iqr)
  ) %>% pivot_longer(-1) %>% pivot_wider(names_from=mdr,values_from=value) %>% 
  separate(4,into=c("C5","C6"), sep='__') %>% 
  separate(3,into=c("C3","C4"), sep='__') %>% 
  separate(2,into=c("C1","C2"), sep='__') %>% 
  mutate(name=sapply(name,function(x)var_label(cost_cat)[[x]])) %>% 
  rename("Cost category"=1,"Mean (SD)"=2,"Median (IQR)"=3,"Mean (SD)"=4,"Median (IQR)"=5,"Mean (SD)"=6,"Median (IQR)"=7)

kable(Table5a_hc, "html", caption = "Table 5. Estimated costs borne by households affected by TB (human capital approach)",
      align = 'rcccccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  #  pack_rows("Hours lost by patient",1,3) %>% 
  add_header_above(c(" " = 1, "DS-TB" = 2, "DR-TB" = 2, "Total" = 2)) %>% 
  pack_rows("Pre-diagnosis",1,3) %>% 
  pack_rows("Post-diagnosis",4,9) %>% 
  pack_rows("Subtotal",10,12) %>%
  pack_rows("Total",13,13) %T>%
  cat(., file = paste0(output_loc,"Table 5a. Estimated costs borne by households affected by TB (HC).html"))

Table5b_op <- 
  rbind(
    cost_cat_op %>% group_by(mdr) %>% summarize_all(mean_sd_med_iqr),
    cost_cat_op %>% group_by(mdr="Total") %>% summarize_all(mean_sd_med_iqr)
  ) %>% pivot_longer(-1) %>% pivot_wider(names_from=mdr,values_from=value) %>% 
  separate(4,into=c("C5","C6"), sep='__') %>% 
  separate(3,into=c("C3","C4"), sep='__') %>% 
  separate(2,into=c("C1","C2"), sep='__') %>% 
  mutate(name=sapply(name,function(x)var_label(cost_cat)[[x]])) %>% 
  rename("Cost category"=1,"Mean (SD)"=2,"Median (IQR)"=3,"Mean (SD)"=4,"Median (IQR)"=5,"Mean (SD)"=6,"Median (IQR)"=7)

kable(Table5b_op, "html", caption = "Table 5b. Estimated costs borne by households affected by TB (output approach)",
      align = 'rcccccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  #  pack_rows("Hours lost by patient",1,3) %>% 
  add_header_above(c(" " = 1, "DS-TB" = 2, "DR-TB" = 2, "Total" = 2)) %>% 
  pack_rows("Pre-diagnosis",1,2) %>% 
  pack_rows("Post-diagnosis",3,7) %>% 
  pack_rows("Subtotal",8,10) %>%
  pack_rows("Total",11,11) %T>%
  cat(., file = paste0(output_loc,"Table 5b. Estimated costs borne by households affected by TB (OP).html"))

```
\     
  
***
\pagebreak
\    
  
#### Figure 5. Breakdown by cost category 
```{r Figure 5, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}

temp <- cost_cat %>% 
  select(mdr,cat_medical,cat_nmedical,cat_indirect,total_cost_hc)
cost_brkdown_hc <- 
  rbind(
    temp %>% group_by(mdr) %>% summarize_all(mean),
    temp %>% group_by(mdr="Total") %>% summarize_all(mean)
  ) %>% pivot_longer(-1) #%>% pivot_wider(names_from=mdr,values_from=value)

cost_brkdown_hc %>% filter(name!="total_cost_hc") %>% 
  mutate(Category=factor(name,levels=c('cat_indirect','cat_nmedical','cat_medical'),
                         labels=c("Indirect","Non-medical","Medical"))) %>% 
  mutate(mdr=factor(mdr,levels=c("DR-TB","DS-TB","Total"),labels=c("DR-TB","DS-TB","Overall"))) %>% 
  ggplot() + 
  geom_bar(aes(fill=Category,y=value,x=mdr),position='stack',stat='identity')+
  scale_fill_discrete(name = NULL,
                    labels = c("Indirect","Direct, non-medical","Direct, medical"),
                    guide = guide_legend(reverse=TRUE))+
  theme(legend.position='top') +
  coord_flip() +
  labs(title="Breakdown by cost category (human capital approach for indirect cost)") +
  xlab("") + ylab("Cost (TZS)")


temp <- cost_cat %>% 
  select(mdr,cat_medical,cat_nmedical,income_diff,total_cost_output)
cost_brkdown_op <- 
  rbind(
    temp %>% group_by(mdr) %>% summarize_all(mean),
    temp %>% group_by(mdr="Total") %>% summarize_all(mean)
  ) %>% pivot_longer(-1) #%>% pivot_wider(names_from=mdr,values_from=value)
cost_brkdown_op %>% filter(name!="total_cost_output") %>% 
  mutate(Category=factor(name,levels=c('income_diff','cat_nmedical','cat_medical'),
                         labels=c("Indirect","Non-medical","Medical"))) %>% 
  mutate(mdr=factor(mdr,levels=c("DR-TB","DS-TB","Total"),labels=c("DR-TB","DS-TB","Overall"))) %>% 
  ggplot() + 
  geom_bar(aes(fill=Category,y=value,x=mdr),position='stack',stat='identity')+
  scale_fill_discrete(name = NULL,
                    labels = c("Indirect","Direct, non-medical","Direct, medical"),
                    guide = guide_legend(reverse=TRUE))+
  theme(legend.position='top') +
  coord_flip() +
  labs(title="Breakdown by cost category (output approach for indirect cost)") +
  xlab("") + ylab("Cost (TZS)")

```
\     
  
***
\pagebreak
\    
  
#### Table 6. Reported dissaving mechanisms and social consequences  
```{r Table 6, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 6. Reported coping mechanisms and social consequences ####
## Tabulation by using 
##  dplyr
##
coping_quint <- d %>% 
  mutate(impact1=as.numeric(cope_impact=='cope_impact1'),impact2=as.numeric(cope_impact=='cope_impact2'),
         impact3=as.numeric(cope_impact=='cope_impact3'),impact4=as.numeric(cope_impact=='cope_impact4')) %>% 
  mutate(days_lost=replace_na(days_lost,0)) %>% 
  select(exp_quintile,borrow,assetsale,coping,
         social_effectsocial_food_ins,social_effectsocial_divorce,social_effectsocial_lossofjo,
         social_effectsocial_dropout,social_effectsocial_exclusio,
         days_lost,impact4,impact3,impact2,impact1) 

var_label(coping_quint) <- list(
  borrow = "Borrowing",
  assetsale = "Sale of assets",
  coping = "Any coping",
  social_effectsocial_food_ins = "Food insecurity",
  social_effectsocial_divorce = "Divorce",
  social_effectsocial_lossofjo = "Loss of job",
  social_effectsocial_dropout = "Child intruputed schooling",
  social_effectsocial_exclusio = "Social exclusion",
  days_lost = "Any days of work lost",
  impact4 = "Much poorer",
  impact3 = "Poorer",
  impact2 = "Unchanged",
  impact1 = "Richer")

Table6 <- 
  rbind(
    coping_quint %>% group_by(exp_quintile) %>% summarise_all(n_perc),  
    coping_quint %>% group_by(exp_quintile="Total") %>% summarise_all(n_perc)
  ) %>% 
  pivot_longer(-1) %>% pivot_wider(names_from=exp_quintile,values_from=value) %>% 
  mutate(name=sapply(name,function(x)var_label(coping_quint)[[x]])) %>% 
  rename(" "=1,"n (%)"=2,"n (%)"=3,"n (%)"=4,"n (%)"=5,"n (%)"=6,"n (%)"=7)

kable(Table6, "html", #caption = "Table 6. Reported dissaving mechanisms and social consequences",
      align = 'rcccccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  #  pack_rows("Hours lost by patient",1,3) %>% 
  add_header_above(c(" "=1,"Poorest"=1,"Second"=1,"Third"=1,"Fourth"=1,"Wealthiest"=1,"Total"=1)) %>% 
  add_header_above(c(" " = 1, "Expenditure Quintile" = 5, " "=1)) %>% 
  pack_rows("Coping strategies",1,3) %>% 
  pack_rows("Social consequences",4,9) %>% 
  pack_rows("Perceived impact",10,13) %T>%
  #  pack_rows("Total",12,12) %T>%
  cat(., file = paste0(output_loc,"Table 6. Reported dissaving mechanisms and social consequences_by_exp_quintile.html"))

```
\     
  
***
\pagebreak
\    
  
#### Table 7. Household classified as facing catastrophic costs by different threshold   
```{r Table 7, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 7. Household classified as facing catastrophic costs by different threshold   ####
## Tabulation by using 
##  dplyr
##

Table7a_1 <- 
  d %>% select(pct1) %>% mutate(pct1_c=cut(pct1,breaks=c(0:6/10,Inf),right=FALSE)) %>% 
  group_by(pct1_c) %>%
  summarize_all(length) %>% rename(thr=1,n=2) %>% arrange(desc(thr)) %>% 
  mutate(cum=cumsum(n),p=cum/sum(n)*100) %>% 
  mutate(np=paste0(cum," (",round(p,1),")")) %>% 
  arrange(thr) %>% slice(-1) %>% mutate(thr=paste0(1:6*10,"%")) %>% select(thr,np)

Table7a_2 <- 
  d %>% select(pct_conservative) %>% mutate(pct_conservative_c=cut(pct_conservative,breaks=c(0:6/10,Inf),right=FALSE)) %>% 
  group_by(pct_conservative_c) %>%
  summarize_all(length) %>% rename(thr=1,n=2) %>% arrange(desc(thr)) %>% 
  mutate(cum=cumsum(n),p=cum/sum(n)*100) %>% 
  mutate(np=paste0(cum," (",round(p,1),")")) %>% 
  arrange(thr) %>% slice(-1) %>% mutate(thr=paste0(1:6*10,"%")) %>% select(thr,np)

Table7a <- rbind(Table7a_1,Table7a_2) %>% 
  mutate(method=rep("",12)) %>% 
  select(method,thr,np) %>% 
  rename(" "=1,"Threshold as % of annual household income"=2,"Households facing catastrophic costs n (%)"=3)

kable(Table7a, "html", caption = "Table 7(a). Household classified as facing catastrophic costs by different threshold (output approach)",
      align = 'cc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  row_spec(2, bold = T, background = "#8EE5EE") %>% 
  pack_rows("Direct and indrect costs (The End TB Strategy indicator)",1,6) %>% 
  pack_rows("Direct costs only (a conservative indicator)",7,12) %T>% 
  cat(., file = paste0(output_loc,"Table 7(a). Household classified as facing catastrophic costs by different threshold(output).html"))
#
#
#
Table7b_1 <- 
  d %>% select(pct3) %>% mutate(pct3_c=cut(pct3,breaks=c(0:6/10,Inf),right=FALSE)) %>% 
  group_by(pct3_c) %>%
  summarize_all(length) %>% rename(thr=1,n=2) %>% arrange(desc(thr)) %>% 
  mutate(cum=cumsum(n),p=cum/sum(n)*100) %>% 
  mutate(np=paste0(cum," (",round(p,1),")")) %>% 
  arrange(thr) %>% slice(-1) %>% mutate(thr=paste0(1:6*10,"%")) %>% select(thr,np)

Table7b_2 <- 
  d %>% select(pct_conservative) %>% mutate(pct_conservative_c=cut(pct_conservative,breaks=c(0:6/10,Inf),right=FALSE)) %>% 
  group_by(pct_conservative_c) %>%
  summarize_all(length) %>% rename(thr=1,n=2) %>% arrange(desc(thr)) %>% 
  mutate(cum=cumsum(n),p=cum/sum(n)*100) %>% 
  mutate(np=paste0(cum," (",round(p,1),")")) %>% 
  arrange(thr) %>% slice(-1) %>% mutate(thr=paste0(1:6*10,"%")) %>% select(thr,np)

Table7b <- rbind(Table7b_1,Table7b_2) %>% 
  mutate(method=rep("",12)) %>% 
  select(method,thr,np) %>% 
  rename(" "=1,"Threshold as % of annual household income"=2,"Households facing catastrophic costs n (%)"=3)

kable(Table7b, "html", caption = "Table 7(b). Household classified as facing catastrophic costs by different threshold (human capital approach)",
      align = 'cc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  row_spec(2, bold = T, background = "#8EE5EE") %>% 
  pack_rows("Direct and indrect costs (The End TB Strategy indicator)",1,6) %>% 
  pack_rows("Direct costs only (a conservative indicator)",7,12) %T>% 
  cat(., file = paste0(output_loc,"Table 7(b). Household classified as facing catastrophic costs by different threshold(HC).html"))

```
\     
  
***
\pagebreak
\    
  
#### Figure 7(a-d). Proportion of catastrophic costs by threashold   
```{r Figure CC threshold, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Figure CC threshold ####
## Figure. Proportion of households faced catastrophic costs by different threshold
##       
##

# var=d$pct1
# from=0
# to=2
# graph=TRUE

threshold_graph <- function(var, from=0, to=2, threshold=c(1:6/10), graph=TRUE){
  x=seq(from=from,to=to,by=1/100)
  y=NULL
  n=length(var)
  for(i in x){
    y=c(y,sum(var>i)/n)
  }
  thcurve <- tibble(x=x,y=y)
  cc_by_threshold <- thcurve[x %in% threshold,] %>% rename(threshold=1,cc=2) %>% 
  mutate(Threshold = factor(paste0(threshold*100,"%")))
    
  pathdata <- 
    cc_by_threshold %>% mutate(x1=0,y1=cc,x2=threshold,y2=cc,x3=threshold,y3=0) %>% 
    pivot_longer(x1:y3,names_to=c(".value","order"),names_pattern = "(.)(.)") %>% 
    mutate(size=ifelse(threshold==0.2,1.2,0.6))
  
  p <- 
    ggplot(thcurve,aes(x=x,y=y)) + geom_line(size=0.6,col='darkblue') +
    scale_color_brewer(palette = "Spectral",direction = 1) +
    scale_x_continuous(breaks=0:4/2,labels=c("0%","50%","100%","150%","200%"),limits=c(0,2))+
    scale_y_continuous(breaks=0:10/10,labels=paste0(0:10*10,"%"))+
    labs(title="Proportion of households faced catastrophic costs by different threshold",
         x="Threadshold as percentage of annual household income/expenditure/consumption (%)",
         y="Proportion of households faced catastrophic costs (%)") + 
    scale_size_identity() +
    geom_path(data=pathdata,aes(x=x,y=y,col=Threshold,size=size), 
              arrow = arrow(ends = "first", type = "open",angle=15,length=unit(0.1, "inches")))
#  if(graph==TRUE) print(p)
#  print(cc_by_threshold[,c('thr_lab','cc')])
  invisible(list(plot=p, thcurve=thcurve, cc_by_threshold=cc_by_threshold))
}

#gh <- read_dta(file = "../tbpcs_ghana/ghana_pcs_clean.dta")
#mmr <- read_dta(file = "../tbpcs_myanmar/mmr_pcs_clean.dta")
#pct1_mmr <- mmr$pct1[!is.na(mmr$pct1)]
#threshold_graph(gh$pct1, from=0, to=2, graph=TRUE)
#ggsave("Ghana_threshold_pct1.pdf")
#gh %>% select(pct1) %>% 
#  arrange(pct1) %>% summary() 
#  mutate(rn=row_number())
#  ggplot(aes(x=pct1)) + stat_ecdf(geom = "step")

#var_label(d$pct1)
res <- threshold_graph(d$pct1,graph=FALSE)
p <- res$plot + labs(subtitle="(a) Cost as percentage of annual household income (output approach)")
p
ggsave(file=paste0(output_loc,"Fig7-a.Proportion of catastrophic costs by threashold (output-income).pdf"),width=8,height=6)

#var_label(d$pct2)
res <- threshold_graph(d$pct2,graph=FALSE)
p <- res$plot + labs(subtitle="(b) Cost as percentage of annual household income (human capital approach)")
p
ggsave(file=paste0(output_loc,"Fig7-b.Proportion of catastrophic costs by threashold (hc-income).pdf"),width=8,height=6)

#var_label(d$pct3)
res <- threshold_graph(d$pct3,graph=FALSE)
p <- res$plot + labs(subtitle="(c) Cost as percentage of annual household expenditure (human capital approach)")
p
ggsave(file=paste0(output_loc,"Fig7-c.Proportion of catastrophic costs by threashold (hc-expenditure).pdf"),width=8,height=6)

#var_label(d$pct_conservative)
res <- threshold_graph(d$pct_conservative,graph=FALSE)
p <- res$plot + labs(subtitle="(d) Cost as percentage of annual household income (direct costs only)")
p
ggsave(file=paste0(output_loc,"Fig7-d.Proportion of catastrophic costs by threashold (direct costs only).pdf"),width=8,height=6)

```
\     
  
***
\pagebreak
\    
  
#### Table 8. Risk factors of experiencing catastrophic costs 
```{r Table 8, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Table 8. Risk factors of experiencing catastrophic costs ####
## 
##
fullmodel <- glm(cc1_f ~ age_c + sex + mdr + hiv + hh_quintile + prev_hosp + self_admin +insurance + 
                   district+ facility_type +delay +ur_ru, data=d, family=binomial())
k_p1=qchisq(0.1,1,lower.tail=FALSE) # define k value corresponding to p=0.05 in Chi-square distribution (one side)
invisible(sw_backward <- step(fullmodel,k=k_p1,trace = 0)) # backward selection from the full model

dependent = 'cc1_f'
explanatory       = c('age_c','sex','mdr','hiv','exp_quintile_r','prev_hosp','self_admin','insurance','delay','ur_ru')
explanatory_final = c('age_c','sex','mdr',      'exp_quintile_r','prev_hosp','self_admin','ur_ru')
t_or <- d %>% 
  finalfit(dependent, explanatory, explanatory_final, digits=c(2,2,3),keep_fit_id = FALSE) %>% 
  ff_remove_p() #%>% fit2df()
t_bivar <- d %>%
  summary_factorlist(dependent, explanatory,p=TRUE, column=FALSE,total_col = TRUE, fit_id=FALSE)

#t_ffmerged <- ff_merge(t_bivar,t_or,last_merge = TRUE) does not work
Table8 <- cbind(t_bivar[,c(1,2,5,4,6)],t_or[,c(5,6)]) %>% remove_rownames() %>% 
  rename("Variable"=1," "=2,"N"=3,"n (%)"=4)
kable(Table8, "html", caption = "Table 8. Risk factors for TB-affected household faceing catastrophic costs (cc1)",
      align = 'llccccc') %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %T>%
  cat(., file = paste0(output_loc,"Table 8. Risk factors for catastrophic costs.html"))

##
## riskfactors for cc3
##
fullmodel <- glm(cc3_f ~ age_c + sex + mdr + hiv + exp_quintile + prev_hosp + self_admin +insurance + 
                   district+ facility_type +delay +ur_ru, data=d, family=binomial())
k_p1=qchisq(0.1,1,lower.tail=FALSE) # define k value corresponding to p=0.05 in Chi-square distribution (one side)
invisible(sw_backward <- step(fullmodel,k=k_p1,trace = 0)) # backward selection from the full model
#step(fullmodel,k=k_p1)
     
dependent = 'cc3_f'
explanatory       = c('age_c','sex','mdr','hiv','exp_quintile_r','prev_hosp','self_admin','insurance','delay','ur_ru')
explanatory_final = c('age_c','sex','mdr',      'exp_quintile_r','prev_hosp','self_admin','ur_ru')
t_or <- d %>% 
  finalfit(dependent, explanatory, explanatory_final, digits=c(2,2,3),keep_fit_id = FALSE) %>% 
  ff_remove_p() #%>% fit2df()
t_bivar <- d %>%
  summary_factorlist(dependent, explanatory,p=TRUE, column=FALSE,total_col = TRUE, fit_id=FALSE)

#t_ffmerged <- ff_merge(t_bivar,t_or,last_merge = TRUE) does not work
Table8_cc3 <- cbind(t_bivar[,c(1,2,5,4,6)],t_or[,c(5,6)]) %>% remove_rownames() %>% 
  rename("Variable"=1," "=2,"N"=3,"n (%)"=4)
kable(Table8_cc3, "html", caption = "Table 8. Risk factors for TB-affected household faceing catastrophic costs (cc3)",
      align = 'llccccc') %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %T>%
  cat(., file = paste0(output_loc,"Table 8. Risk factors for catastrophic costs (cc3).html"))

```
\     
  
***
\pagebreak
\    
  
#### Figure 9. Distribution of income/expenditure measurements   
```{r Figure income_expenditure, echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## Figure distribution of income and expenditure ####
## 
##       
##
ggplot(d) +
  geom_histogram(aes(x=income_hh_pre_annual/1000,stat(density)),
                 col='grey',fill='skyblue',alpha=0.7,binwidth = 1000) +
  geom_density(aes(x=income_hh_pre_annual/1000),col='lightblue',fill='lightcyan',alpha=0.4)+
  xlim(0,40000) + ylim(0,0.00028) +
  labs(title="Distribution of annual household income before TB diagnosis") +
  xlab("Annual household income (1000 TZS)") + ylab("Density")
ggsave("Fig.Desity_income_hh_pre.pdf",width=9,height=6)

ggplot(d) +
  geom_histogram(aes(x=expend_hh_annual/1000,y=..density..),
                 col='grey',fill='pink',alpha=0.7,binwidth = 1000) +
  geom_density(aes(expend_hh_annual/1000),col='lightyellow',fill='pink',alpha=0.4)+
  xlim(0,40000) + ylim(0,0.00028) +
  labs(title="Distribution of annual household expenditure") +
  xlab("Annual household expenditure (1000 TZS)") + ylab("Density")
ggsave("Fig.Desity_expend_hh_annual.pdf",width=9,height=6)

ggplot(d) +
  geom_density(aes(x=expend_hh_annual/1000,col='a',fill='a'),alpha=0.4)+
  geom_density(aes(x=income_hh_pre_annual/1000,col='b',fill='b'),alpha=0.4)+
  xlim(0,40000) + ylim(0,0.00028) +
  labs(title="Distribution of annual household income and expenditure") +
  xlab("Annual household income and expenditure (1000 TZS)") + ylab("Density")+
  scale_color_manual(guide = "legend", name = " ",
                       values = c('a'='red', 'b'='blue'),
                       labels = c("Expenditure","Income")) +
  scale_fill_manual(guide = "legend",name = " ",
                       values = c('a'='red', 'b'='blue'),
                       labels = c("Expenditure","Income")) 
ggsave("Fig.Desity_income_vs_expenditure.pdf",width=9,height=6)

ggplot(d) +
  geom_density(aes(x=expend_hh_annual/1000,col='a',fill='a'),alpha=0.4)+
  geom_density(aes(x=income_hh_pre_annual_reported/1000,col='b',fill='b'),alpha=0.4)+
  xlim(0,40000) + ylim(0,0.00028) +
  labs(title="Distribution of annual household income and expenditure by urban/rural") +
  xlab("Annual household income and expenditure (1000 TZS)") + ylab("Density")+
  scale_color_manual(guide = "legend", name = " ",
                       values = c('a'='red', 'b'='blue'),
                       labels = c("Expenditure","Income")) +
  scale_fill_manual(guide = "legend",name = " ",
                       values = c('a'='red', 'b'='blue'),
                       labels = c("Expenditure","Income")) +
  facet_wrap(~ur_ru,ncol=1)
ggsave("Fig.Desity_income_vs_expenditure_byUrbanRural.pdf",width=9,height=6)

ggplot(d,aes(x=income_hh_pre_annual/1000,y=expend_hh_annual/1000))+
  geom_abline(slope=1,intercept = 0,col='white',size=1)+
  geom_point(aes(color=ur_ru),alpha=0.7) +
  geom_smooth(method = "lm") +
  xlim(0,40000) + ylim(0,40000)+
#  scale_x_log10() + scale_y_log10() +
  labs(title="Correlation between annual household income and expenditure") +
  xlab("Annual household income (1000 TZS)") + ylab("Annual household expenditure (1000 TZS)") +
  scale_color_discrete(name="Urban/Rural",labels=c("Rural","Semi-urban","Urban"))
ggsave("Fig.Scatter_income_vs_expenditure.pdf",width=9,height=6)


```
\     
  
***
\pagebreak
\    
#### Adjustment for survey design (using cc1)
```{r adjustment for survey design (cc1), echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}
## svydesing settings #######
## adjustment 
## 
## survey design setting 
## 
## srs / cluster_design / cluster_design_wt

# assuming simple random sampling (theoritically results should be same as above)
srs <- svydesign(id=~1, data=d, probs = 1) ## assuming simple random sampling 
# setting cluster sampling design (district as cluster)
cluster_design <- svydesign(id=~district, data=d, probs = 1) # set district as cluster sampling unit 
# setting cluster sampling design (district as cluster) and weighted for non-response 
#All TB cases notified in 2018: 75828
#DR-TB Patients started on treatment in 2018: 409
#DS-TB in 2018: 75828-409=75419
# DR-TB: 0.539%  DS-TB: 99.461%
tball <- 75828
csize <- 26
d <- 
  d %>% 
  group_by(district) %>% summarize(actualsize=n()) %>% # actual cluster size
  mutate(nc = round(actualsize/csize,0))     %>%       # 
  mutate(dc = csize*nc)                      %>%       # design-based cluster size 
  mutate(sprov_des = csize/tball)            %>%       # (1) sampling probability (design-based)
  mutate(sprov_enr = actualsize/(csize*nc))  %>%      # (2) adjust probability by enrolment 
  mutate(sampprov  = actualsize/(tball*nc))  %>%       # (3) sampling probability (overall)
  mutate(wt = 1/sampprov) %>% select(district,sampprov,wt) %>% # create data.frame with district and weight
  left_join(d,.) # join with original d 
d %>% ggplot(aes(x=sampprov)) + geom_histogram() + #check distribution of sampling probabilities 
  labs(title="Distribution of sampling probability") + xlab("Sampling probability") + ylab("Number of sample")
ggsave(file=paste0(output_loc,"Distribution of sampling weight.pdf"),width=6,height=4)

# setting cluster sampling design (district as cluster) with wt for non-response adjustment (under- or over-enrolment)
cluster_design_wt <- svydesign(id=~district, data=d, weights=~wt)

cc1_by_design <- 
  rbind(
    rbind(
      svyby(~cc1,~mdr,   srs,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc1,~all__1,srs,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("SRS")),
    rbind(
      svyby(~cc1,~mdr,   cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc1,~all__1,cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("Cluster")),
    rbind(
      svyby(~cc1,~mdr,   cluster_design_wt,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc1,~all__1,cluster_design_wt,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("Weighted"))
  ) %>% add_column(grp=rep(c("DS-TB","DR-TB","Total"),3),.before=1) %>% as_tibble() %>% 
  mutate(grp=factor(grp),design=factor(design,levels=c("Weighted","Cluster","SRS")))

cc1_by_design %>% 
  mutate(p_ci=paste0(round(cc1,1),"% (",round(ci_l,1),"-",round(ci_u,1),")")) %>% 
  select(grp,p_ci,design) %>% 
  pivot_wider(names_from = grp, values_from=p_ci) %>% 
  rename("Design assumption"=1) ->cc1_by_design_w
# cc1_by_design_w %>% 
#   kable("html", caption = "Table #. Effect of analysis taking into account sampling design",
#         align = 'rccc', na_include = TRUE) %>%
#   kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
#   add_header_above(c(" " = 1, "Percentage of catastrophic costs (95% CI)" = 3)) %T>% 
# #  pack_rows("Percentage of catastrophic costs (95% CI)",1,3) %T>%
#   cat(., file = paste0(output_loc,"Table #. Effect of analysis taking into account sampling design.html"))

p <- 
  ggplot(cc1_by_design,aes(x=design,y=cc1,col=grp)) +
  geom_point() + 
  geom_errorbar(aes(ymin=ci_l, ymax=ci_u,col=grp))+
  #                width=0.5,cex=1,position=position_dodge(width=0.5))+ 
  facet_wrap(~grp,ncol = 1) +
  scale_y_continuous(limits = c(0,100)) +
  coord_flip() +
  labs(title = "Estiamtes and confidence intervals adjusted by sampling design", subtitle = "Percentage of households experienced catastrophic costs") +
  xlab("Adjusted by sampling design")+
  ylab("Proportion (%)")
p
ggsave(paste0(output_loc,"Fig.Effect of sampling design (output-income).pdf"),plot=p, width=6, height=4)

#
# formula for deff, rho, k and m
#
# deff = (1+(m-1)*k^2*p/(1-p))
# deff = 1+(m-1)*rho
# rho = (deff-1)/(m-1)
# k = sqrt(rho*(1-p)/p)

m = 26 # cluster size 
rbind(
  #  svymean(~cc1,srs,deff="replace") %>% as.data.frame() %>% select(1,3),
  svymean(~cc1,cluster_design,deff="replace") %>% as.data.frame() %>% select(1,3),
  svymean(~cc1,cluster_design_wt,deff="replace") %>% as.data.frame() %>% select(1,3)
) %>% 
  rename("p"=1) %>% 
  mutate(rho=(deff-1)/(m-1) %>% round(digits=4)) %>%
  mutate(k=sqrt(rho*(1-p)/p)) -> obs_deff

obs_deff %>% mutate(DEFF=as.character(round(deff,digits=3)),Rho=as.character(round(rho,digits=5)),
                    k=as.character(round(k,5))) %>% select(DEFF,Rho,k) %>% 
  add_row(DEFF=1,Rho="—",k="—",.before = 1) %>% bind_cols(cc1_by_design_w,.) %>% 
  kable("html", #caption = "Table #. Effect of analysis taking into account sampling design and observed design effect paramters",
        align = 'rccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  add_header_above(c(" " = 1, "Percentage of catastrophic costs (95% CI)" = 3,
                     "Observed design effect paramters"=3)) %T>% 
  #  pack_rows("Percentage of catastrophic costs (95% CI)",1,3) %T>%
  cat(., file = paste0(output_loc,"Table #. Effect of analysis taking into account sampling design and observed deff(cc1).html"))

# svy commands examples 
# svytotal(~cc1,cluster_design)
# svytable(~cc1,cluster_design)
# svyciprop(~I(cc1==1), design=cluster_design, method=c('asin'))
# svyby(~cc1,~mdr,cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
# cbind((t <- svymean(~total_cost_hc,cluster_design)),confint(t))
# svyquantile(~total_cost_hc,cluster_design_wt,quantiles=c(0.5,0.25,0.75))
#
#

```
     \     
  
***
\pagebreak
\    
#### Adjustment for survey design (using cc3)
```{r adjustment for survey design (cc3), echo=FALSE, message=FALSE, warning=FALSE, paged.print=FALSE}

##
##
##
cc3_by_design <- 
  rbind(
    rbind(
      svyby(~cc3,~mdr,   srs,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc3,~all__1,srs,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("SRS")),
    rbind(
      svyby(~cc3,~mdr,   cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc3,~all__1,cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("Cluster")),
    rbind(
      svyby(~cc3,~mdr,   cluster_design_wt,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5),
      svyby(~cc3,~all__1,cluster_design_wt,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
    )  %>% add_column(design=c("Weighted"))
  ) %>% add_column(grp=rep(c("DS-TB","DR-TB","Total"),3),.before=1) %>% as_tibble() %>% 
  mutate(grp=factor(grp),design=factor(design,levels=c("Weighted","Cluster","SRS")))

cc3_by_design %>% 
  mutate(p_ci=paste0(round(cc3,1),"% (",round(ci_l,1),"-",round(ci_u,1),")")) %>% 
  select(grp,p_ci,design) %>% 
  pivot_wider(names_from = grp, values_from=p_ci) %>% 
  rename("Design assumption"=1) ->cc3_by_design_w
# cc1_by_design_w %>% 
#   kable("html", caption = "Table #. Effect of analysis taking into account sampling design",
#         align = 'rccc', na_include = TRUE) %>%
#   kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
#   add_header_above(c(" " = 1, "Percentage of catastrophic costs (95% CI)" = 3)) %T>% 
# #  pack_rows("Percentage of catastrophic costs (95% CI)",1,3) %T>%
#   cat(., file = paste0(output_loc,"Table #. Effect of analysis taking into account sampling design.html"))

p <- 
  ggplot(cc3_by_design,aes(x=design,y=cc3,col=grp)) +
  geom_point() + 
  geom_errorbar(aes(ymin=ci_l, ymax=ci_u,col=grp))+
  #                width=0.5,cex=1,position=position_dodge(width=0.5))+ 
  facet_wrap(~grp,ncol = 1) +
  scale_y_continuous(limits = c(0,100)) +
  coord_flip() +
  labs(title = "Estiamtes and confidence intervals adjusted by sampling design", subtitle = "Percentage of households experienced catastrophic costs") +
  xlab("Adjusted by sampling design")+
  ylab("Proportion (%)")
p
ggsave(paste0(output_loc,"Fig.Effect of sampling design (hc-expenditure).pdf"),plot=p, width=6, height=4)

#
# formula for deff, rho, k and m
#
# deff = (1+(m-1)*k^2*p/(1-p))
# deff = 1+(m-1)*rho
# rho = (deff-1)/(m-1)
# k = sqrt(rho*(1-p)/p)


m = 26 # cluster size 
rbind(
  #  svymean(~cc1,srs,deff="replace") %>% as.data.frame() %>% select(1,3),
  svymean(~cc3,cluster_design,deff="replace") %>% as.data.frame() %>% select(1,3),
  svymean(~cc3,cluster_design_wt,deff="replace") %>% as.data.frame() %>% select(1,3)
) %>% 
  rename("p"=1) %>% 
  mutate(rho=(deff-1)/(m-1) %>% round(digits=4)) %>%
  mutate(k=sqrt(rho*(1-p)/p)) -> obs_deff

obs_deff %>% mutate(DEFF=as.character(round(deff,digits=3)),Rho=as.character(round(rho,digits=5)),
                    k=as.character(round(k,5))) %>% select(DEFF,Rho,k) %>% 
  add_row(DEFF=1,Rho="—",k="—",.before = 1) %>% bind_cols(cc3_by_design_w,.) %>% 
  kable("html", #caption = "Table #. Effect of analysis taking into account sampling design and observed design effect paramters",
        align = 'rccc', na_include = TRUE) %>%
  kable_styling(bootstrap_options = c("striped"), full_width=F,position='left') %>% 
  add_header_above(c(" " = 1, "Percentage of catastrophic costs (95% CI)" = 3,
                     "Observed design effect paramters"=3)) %T>% 
  #  pack_rows("Percentage of catastrophic costs (95% CI)",1,3) %T>%
  cat(., file = paste0(output_loc,"Table #. Effect of analysis taking into account sampling design and observed deff(cc3).html"))

# svy commands examples 
# svytotal(~cc1,cluster_design)
# svytable(~cc1,cluster_design)
# svyciprop(~I(cc1==1), design=cluster_design, method=c('asin'))
# svyby(~cc1,~mdr,cluster_design,svyciprop,vartype="ci")[,-1] %>% '*'(100) %>% round(digits=5)
# cbind((t <- svymean(~total_cost_hc,cluster_design)),confint(t))
# svyquantile(~total_cost_hc,cluster_design_wt,quantiles=c(0.5,0.25,0.75))
#
#

```
     \     
  
***
\pagebreak
\
